Do function application with a hook
The hook func:compile_function/4 lets others define function behavior
for any term they like.  For example, someone who implements a hash
table might define function application as looking up a key in the
table and throwing an exception if it doesn't exist.

This makes it easier for me to add different kinds of functions.  It
also acknowledges that I won't come up with all the ways function
application might be useful.