Working TAP testing system
To create a test file, one imports library(tap) then defines a bunch
of predicates.  Each predicate is a test case whose name is its
description.  Failing test cases are handled too.

This makes it easy to define test cases with the syntactic overhead of
PlUnit and with all the conveniences of the TAP testing ecosystem.
This TAP code should be moved to a separate pack soon.