Fix function composition across modules
Each composed function is implemented by a predicate which is asserted at
compile time.  We used to add those predicates to the func module.
Unfortunately, that meant that any predicates called inside a
composition were being called from the context of the func module.

By asserting implementation predicates in the module where the
composition occurs, the context of implementation is the same as the one
used during composition.  This assures that predicate resolution works
as intended.

Closes #11